Rename hook to useDataErrorCallback
